feat: refactor bundler
